% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mefd_fun.R
\name{mefd_search}
\alias{mefd_search}
\title{Buscador de indicadores}
\usage{
mefd_search(value, config = conf)
}
\arguments{
\item{value}{palabra a buscar en la lista de indicadores}

\item{config}{lista con parámetros de configuración (default)}
}
\value{
data.frame con el nombre y idserie de los indicadores
}
\description{
Busca en archivo metadatos el nombre de indicadores que contienen una palabra
}
\examples{
# Indicadores que contienen la palabra "idoneidad"
mefd_search("idoneidad")
# Indicadores que contienen la palabra "primaria"
mefd_search("primaria")
# Indicadores que contienen las palabra "primaria" y "sexo
mefd_search("primaria.*sexo")
# Indicadores que contienen las palabra "extranjero" o "idoneidad"
mefd_search("extranjero|idoneidad")
}
