% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\docType{data}
\name{meta_mefd}
\alias{meta_mefd}
\title{meta_mefd}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 271 rows and 8 columns.
}
\usage{
meta_mefd
}
\description{
Metadatos de indicadores. Incluye el nombre del indicadore, el archivo .csv, el url.
}
\keyword{datasets}
